refactor: 重构v0.2.0版本
