+++ internal/cache/byteview.go
