+++ internal/store/hash.go
