+++ internal/store/lru2_test.go
