+++ internal/store/lru_test.go
