--- proto/kama/v1/kama.proto
@@ -1,34 +1,0 @@
-syntax = "proto3";
-package kama.v1;
-
-option go_package = "/kamapb";
-
-service KamaCacheService {
-  rpc Get(GetRequest) returns (GetResponse);
-  rpc Set(SetRequest) returns (SetResponse);
-  rpc Delete(DeleteRequest) returns (DeleteResponse);
-}
-
-message GetRequest {
-  string group = 1;
-  string key = 2;
-}
-message GetResponse {
-  bytes value = 1;
-}
-message SetRequest {
-  string group = 1;
-  string key = 2;
-  bytes value = 3;
-}
-message SetResponse {
-  bytes value = 1;
-}
-message DeleteRequest {
-  string group = 1;
-  string key = 2;
-}
-
-message DeleteResponse {
-  bool value = 1;
-}
